Move bench package out of deploy package
